Section 3 - CloudBuild-CICD/cloudrun-ml-models/bikeshare-model/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
scikit-learn==1.2.2 #OPTIONAL
xgboost==1.6.2
pandas==2.0.3
pyarrow==15.0.0
pytest==8.3.1
category_encoders
requests==2.31.0
google-cloud-storage
Werkzeug==2.0.3
numpy==1.24.4
joblib==1.3.2